pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "research-lab-bio"
version = "0.1.0"
description = "A link-in-bio style public landing page with rich content and administration."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flet>=0.25.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0",
    "argon2-cffi>=23.1.0",
    "matplotlib>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-json-report>=1.5.0",
    "ruff>=0.8.0",
    "mypy>=1.6.0",
    "types-PyYAML",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
